Add files via upload
Bhybridnum.pl is a Perl program that takes in an n-by-m binary matrix M representing n taxa and m characters (clusters) and a target, t, and creates
a CNF formula that can be satisfied if and only if there is a Reticulation Network with n leaves, labeled by the taxa, that displays all of the clusters in M,
using at most t reticulation nodes.  That CNF formula can then be passed to a SAT solver to determine if it is satisfiable or not. If satisfiable, the variables set true
specify the nodes and edges in the Reticulation Network that displays the clusters, using at most t reticulation nodes. The program also produces
an ILP formulation whose solution  details a Reticulation Network for the clusters in M, minimizing the number of reticulation nodes used. The ILP is more limited
in the range of data (input matrices) that it can effectively solve.